# specs/001-auth-customer-staff/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dahab Auth API
  version: 1.0.0
  description: |
    Authentication surface for the Dahab Gold Marketplace Backend.
    Two principals — customer and dashboard staff — share primitives
    (rate limits, audit log, error envelope) but do not share tokens.
    All rules trace back to `docs/Technical Spec/dahab-spec-part1-auth.md`.

    Two API surfaces, two guards, four token abilities:

    | Surface | Prefix | Guard | Access ability | Refresh ability |
    |---|---|---|---|---|
    | Customer API | `/customer/*` | `auth:customer` | `customer:access` | `customer:refresh` |
    | Dashboard API | `/dashboard/*` | `auth:staff` | `staff:access` | `staff:refresh` |

    A token of the other principal is rejected with 401 `unauthenticated`. A token
    of the right principal but the wrong ability (a refresh token on an access
    endpoint, or the reverse) is rejected with 403 `forbidden`.

    `x-implemented: true` marks operations that exist in the running API today;
    `x-implemented: false` marks planned operations (see `../tasks.md`).
servers:
  - url: /api/v1
    description: v1
components:
  securitySchemes:
    customerBearer:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: Customer API. The customer access token (ability `customer:access`).
    customerRefreshBearer:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: Customer API refresh endpoint only. The customer refresh token (ability `customer:refresh`).
    dashboardBearer:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: Dashboard API. The staff access token (ability `staff:access`).
    dashboardRefreshBearer:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: Dashboard API refresh endpoint only. The staff refresh token (ability `staff:refresh`).
  schemas:
    Envelope:
      type: object
      required: [data]
      properties:
        data:
          nullable: true
        meta:
          type: object
...
    Session:
      type: object
      required: [access_token, access_token_expires_at, refresh_token, refresh_token_expires_at, token_type]
      properties:
        token_type: { type: string, enum: [Bearer] }
        access_token: { type: string, description: "Sanctum PAT, ~15 min TTL, ability `<principal>:access`" }
        access_token_expires_at: { type: string, format: date-time }
        refresh_token: { type: string, description: "Sanctum PAT with ability `<principal>:refresh`, 30 day TTL, rotates on every use" }
        refresh_token_expires_at: { type: string, format: date-time }
        family_id: { type: string, format: uuid }
...
    MfaChallenge:
      type: object
      required: [mfa_required, session_ref, expires_at]
      properties:
        mfa_required: { type: boolean, enum: [true] }
        session_ref: { type: string, format: uuid }
        expires_at: { type: string, format: date-time }
    MfaEnrollmentPending:
      type: object
      required: [mfa_enrollment_required, otpauth_url, recovery_codes, session_ref]
      properties:
        mfa_enrollment_required: { type: boolean, enum: [true] }
        otpauth_url: { type: string }
        recovery_codes:
          type: array
          items: { type: string }
        session_ref: { type: string, format: uuid }
...
    Staff:
      type: object
      required: [id, role, full_name, email, is_active, permissions]
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [ceo, coo, finance, operations, verification, igi_branch] }
        full_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        is_active: { type: boolean }
        branch_id: { type: integer, nullable: true }
        roles:
          type: array
          description: Spatie roles held on the staff guard
          items: { type: string }
          example: [ceo]
        permissions:
          type: array
          description: Effective Spatie permissions (direct + via roles)
          items: { type: string }
          example: [customer.suspend]
        mfa_enrolled: { type: boolean }
...
paths:
...
  /dashboard/auth/login:
    post:
      tags: [Dashboard Auth]
      x-implemented: false
      security: []
      summary: Staff sign-in (email + password; TOTP for ceo/coo/finance)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, writeOnly: true }
      responses:
        '200':
          description: Session issued OR MFA challenge OR MFA enrollment required
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: '#/components/schemas/Envelope'
                      - type: object
                        properties:
                          data:
                            type: object
                            properties:
                              staff: { $ref: '#/components/schemas/Staff' }
                              session: { $ref: '#/components/schemas/Session' }
                  - allOf:
                      - $ref: '#/components/schemas/Envelope'
                      - type: object
                        properties:
                          data: { $ref: '#/components/schemas/MfaChallenge' }
                  - allOf:
                      - $ref: '#/components/schemas/Envelope'
                      - type: object
                        properties:
                          data: { $ref: '#/components/schemas/MfaEnrollmentPending' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '403': { $ref: '#/components/responses/AccountFrozen' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /dashboard/auth/mfa/verify:
    post:
      tags: [Dashboard Auth]
      x-implemented: false
      security: []
      summary: Complete the TOTP challenge for a pending staff sign-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_ref, code]
              properties:
                session_ref: { type: string, format: uuid }
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          staff: { $ref: '#/components/schemas/Staff' }
                          session: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/MfaInvalid' }
  /dashboard/auth/mfa/enroll:
    post:
      tags: [Dashboard Auth]
      x-implemented: false
      security: []
      summary: Complete MFA enrollment by confirming a TOTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_ref, code]
              properties:
                session_ref: { type: string, format: uuid }
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: MFA enrolled; session issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          staff: { $ref: '#/components/schemas/Staff' }
                          session: { $ref: '#/components/schemas/Session' }
                          recovery_codes:
                            type: array
                            items: { type: string }
        '401': { $ref: '#/components/responses/MfaInvalid' }
  /dashboard/auth/refresh:
    post:
      tags: [Dashboard Auth]
      x-implemented: true
      security:
        - dashboardRefreshBearer: []
      summary: Rotate the staff refresh token
      description: The bearer token here MUST be a staff refresh token (ability `staff:refresh`); an access token is 403. Replaying an already-rotated refresh token revokes the whole family.
      responses:
        '200':
          description: New access/refresh pair, same family
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/RefreshInvalid' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /dashboard/auth/logout:
    post:
      tags: [Dashboard Auth]
      x-implemented: false
      security:
        - dashboardBearer: []
      summary: Revoke the current staff session
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /dashboard/auth/logout-all:
    post:
      tags: [Dashboard Auth]
      x-implemented: false
      security:
        - dashboardBearer: []
      summary: Revoke every staff token for the current staff
      responses:
        '204': { description: Logged out from all devices }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /dashboard/auth/me:
    get:
      tags: [Dashboard Auth]
      x-implemented: true
      security:
        - dashboardBearer: []
      summary: Get the authenticated staff and effective permission set
      responses:
        '200':
          description: Own staff row + permissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Staff' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...